add retry logic to model downloading

+++ onnx/backend/test/runner/__init__.py
@@ -175,20 +175,32 @@ class Runner(object):
             # On Windows, NamedTemporaryFile can not be opened for a
             # second time
             download_file = tempfile.NamedTemporaryFile(delete=False)
-            try:
-                download_file.close()
-                print('Start downloading model {} from {}'.format(
-                    model_test.model_name, model_test.url))
-                urlretrieve(model_test.url, download_file.name)
-                print('Done')
-                with tarfile.open(download_file.name) as t:
-                    t.extractall(models_dir)
-            except Exception as e:
-                print('Failed to prepare data for model {}: {}'.format(
-                    model_test.model_name, e))
-                raise
-            finally:
-                os.remove(download_file.name)
+
+            # Remainning retry times
+            # 0  =>  download failed, raise exception
+            # -1 =>  download succeed
+            retry_times = 3
+
+            while retry_times:
+                try:
+                    download_file.close()
+                    print('Start downloading model {} from {}'.format(
+                        model_test.model_name, model_test.url))
+                    urlretrieve(model_test.url, download_file.name)
+                    print('Done')
+                    with tarfile.open(download_file.name) as t:
+                        t.extractall(models_dir)
+                except Exception as e:
+                    retry_times -= 1
+                    print('Failed to prepare data for model {}: {}, {} retry times remain'.format(
+                        model_test.model_name, e, retry_times))
+                    if retry_times == 0:
+                        raise
+                else:
+                    retry_times = -1
+                finally:
+                    if retry_times <= 0:
+                        os.remove(download_file.name)
         return model_dir
 
     def _add_test(self,
